feat: create initial version of hugo-powered RSS feeds/mirrors
The idea behind this project is to create a Hugo-powered static site
that aggregates/mirrors/proxies one or more RSS feeds.

Basic Workflow
==============

- Import entries from an OPML file,
  creating stub directories for each podcast.
- For each directory:
    - Fetch the updated RSS feed.
    - Save the metadata as a Hugo post.
    - Optionally, save static assets:
        - a copy of the podcast/epidode artwork
    - For each episode:
        - Save the metadata as a Hugo post.
        - Optionally, save static assets:
            - a copy of the podcast/epidode artwork
            - a copy of the episode audio/video
- Build and publish the Hugo site.

This project then distributes the content and static assets
as a Hugo module. The idea being, it should be trivial to build a network
of syndicated content providers, each exposing their content this way,
which would allow redistribution to be as easy adding a new Hugo module
to the project.

Note: In order to be flexible and to support workflows that both save
and do _not_ save copies of the media/static assets, we do a bit of
git-fu and opinionated branch management.

These advanced use-cases should be further documented at a later date.